# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PrimeTrade Task Manager API",
    "version": "1.0.0",
    "description": "JWT-authenticated REST API for task management with RBAC"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin"] }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "description": { "type": "string" },
          "status": { "type": "string", "enum": ["pending", "in-progress", "completed"] },
          "priority": { "type": "string", "enum": ["low", "medium", "high"] },
          "owner": { "$ref": "#/components/schemas/User" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": { "type": "array", "items": { "type": "object" } }
        }
      }
    }
  },
  "tags": [
    { "name": "Auth", "description": "Authentication endpoints" },
    { "name": "Tasks", "description": "Task CRUD endpoints" }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string", "example": "Alice" },
                  "email": { "type": "string", "example": "alice@example.com" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Login and receive JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "alice@example.com" },
                  "password": { "type": "string", "example": "secret123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "Get current user",
        "tags": ["Auth"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Current user data" },
          "401": { "description": "Not authorized" }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "summary": "Get all tasks",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "status", "schema": { "type": "string", "enum": ["pending", "in-progress", "completed"] } },
          { "in": "query", "name": "priority", "schema": { "type": "string", "enum": ["low", "medium", "high"] } },
          { "in": "query", "name": "page", "schema": { "type": "integer", "default": 1 } },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 } }
        ],
        "responses": {
          "200": { "description": "Paginated task list" },
          "401": { "description": "Not authorized" }
        }
      },
      "post": {
        "summary": "Create a task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "status": { "type": "string", "enum": ["pending", "in-progress", "completed"] },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Task created" },
          "400": { "description": "Validation error" }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task data" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "summary": "Update a task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" }, "example": "64abc123def456" }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string", "example": "Updated task title" },
                  "description": { "type": "string", "example": "Updated description" },
                  "status": { "type": "string", "enum": ["pending", "in-progress", "completed"], "example": "in-progress" },
                  "priority": { "type": "string", "enum": ["low", "medium", "high"], "example": "high" }
                }
              },
              "example": {
                "status": "in-progress",
                "priority": "high"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Task updated" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "tags": ["Tasks"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "Task deleted" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}